Create remote asynchronously
There's a lot of changes in this commit because I also
refactored it to build a flat Promise structure instead
a nested one

The core of this commit is that it basically changes
from synchronously calling the
'Git.Remote.create(repository, "origin", serverUrl)
to returning it to the Promise chain.

'Git.Remote.create' returns a Promise which will be resolved
once the remote is created. Since we were calling it
synchronously, it was reaching the last 'then' before
the remote was created.